backend/requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0
transformers==4.35.0
torch==2.0.0
vaderSentiment==3.3.2
numpy<2.0
gunicorn==20.1.0
psycopg2-binary==2.9.9
google-cloud-storage==2.10.0
google-cloud-logging==3.5.0
google-auth==2.23.4
pyjwt==2.8.0
prometheus-client==0.18.0
python-json-logger==2.0.7
redis==5.0.1
huggingface-hub==0.23.0
requests==2.31.0